Перенос моделей-представлений в доменную часть
