+++ Common/WebStore.Domain/ViewModels/OrderViewModel.cs
@@ -1,20 +1,18 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
-    public class UserOrderViewModel
+    public class OrderViewModel
     {
-        public int Id { get; set; }
-        
         [Required]
         public string Name { get; set; }
 
         [Required]
+        [DataType(DataType.PhoneNumber)]
         public string Phone { get; set; }
 
         [Required]
-        public string Address { get; set; }
+        public string Adress { get; set; }
 
-        public decimal TotalSum { get; set; }
     }
 }
