+++ Common/WebStore.Domain/ViewModels/ProductViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class ProductViewModel : INamedEntity, IOrderedEntity
     {
