+++ Common/WebStore.Domain/ViewModels/UserOrderViewModel.cs
@@ -1,0 +1,20 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace WebStore.Domain.ViewModels
+{
+    public class UserOrderViewModel
+    {
+        public int Id { get; set; }
+
+        [Required]
+        public string Name { get; set; }
+
+        [Required]
+        public string Phone { get; set; }
+
+        [Required]
+        public string Address { get; set; }
+
+        public decimal TotalSum { get; set; }
+    }
+}
