+++ UI/WebStore/Infrastructure/Services/InCookies/CookiesCartService.cs
@@ -7,9 +7,9 @@ using System.Net;
 using System.Threading.Tasks;
 using WebStore.Domain;
 using WebStore.Domain.Entities;
+using WebStore.Domain.ViewModels;
 using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Mapping;
-using WebStore.ViewModels;
 
 namespace WebStore.Infrastructure.Services.InCookies
 {
