--- UI/WebStore/ViewModels/OrderViewModel.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace WebStore.ViewModels
-{
-    public class OrderViewModel
-    {
-        [Required]
-        public string Name { get; set; }
-
-        [Required]
-        [DataType(DataType.PhoneNumber)]
-        public string Phone { get; set; }
-
-        [Required]
-        public string Adress { get; set; }
-
-    }
-}
